pyproject.toml
[tool.poetry]
name = "olric-bookstore"
version = "1.0.0"
description = "Online bookstore platform for selling books in Azerbaijan"
authors = ["Olric Team <team@olricbookstore.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "apps"}, {include = "config"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
djangorestframework = "^3.14.0"
pillow = "^10.1.0"
mysqlclient = "^2.2.0"
python-decouple = "^3.8"
django-cors-headers = "^4.3.1"
django-ratelimit = "^4.1.0"
django-axes = "^6.1.1"
requests = "^2.31.0"
gunicorn = "^21.2.0"
django-widget-tweaks = "^1.5.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.7.0"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"
faker = "^20.1.0"
black = "^23.12.1"
flake8 = "^7.0.0"
isort = "^5.13.2"
django-debug-toolbar = "^4.2.0"
locust = "^2.20.0"
safety = "^2.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | migrations
  | staticfiles
  | media
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_glob = ["*/migrations/*", "*/staticfiles/*", "*/media/*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.testing"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=apps",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--verbose",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["apps"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/admin.py",
    "*/apps.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
